(#522) Preprocessor: adjust namespaces, cleanup
