+++ Cesium.Preprocessor/ConditionExpressions/DefinedExpression.cs
