--- Cesium.Preprocessor/ConditionalElementResult.cs
@@ -1,3 +1,0 @@
-namespace Cesium.Preprocessor;
-
-public record ConditionalElementResult(string KeyWord, bool Flag, bool? UpperFlag);
